test/fixtures/wpt/encoding/sharedarraybuffer.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

test(() => {
  const decoder = new TextDecoder('utf-8');
  const buffer = new SharedArrayBuffer(4);
  assert_equals(decoder.decode(new Uint8Array(buffer)), "\u0000\u0000\u0000\u0000");
}, 'decoding SharedArrayBuffer');

</script>
